+++ src/main/java/com/goodee/corpdesk/attendance/repository/AttendanceRepository.java
@@ -1,7 +1,8 @@
-package com.goodee.corpdesk.attendance;
+package com.goodee.corpdesk.attendance.repository;
 
 import java.util.List;
 
+import com.goodee.corpdesk.attendance.entity.Attendance;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.NativeQuery;
