+++ src/main/java/com/goodee/corpdesk/attendance/service/AttendanceService.java
@@ -1,13 +1,14 @@
-package com.goodee.corpdesk.attendance;
+package com.goodee.corpdesk.attendance.service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import com.goodee.corpdesk.attendance.DTO.ResAttendanceDTO;
+import com.goodee.corpdesk.attendance.entity.Attendance;
+import com.goodee.corpdesk.attendance.repository.AttendanceRepository;
 import com.goodee.corpdesk.vacation.entity.VacationDetail;
 import com.goodee.corpdesk.vacation.repository.VacationDetailRepository;
-import com.goodee.corpdesk.vacation.repository.VacationRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
