+++ src/main/java/com/goodee/corpdesk/employee/EmployeeController.java
@@ -2,12 +2,15 @@ package com.goodee.corpdesk.employee;
 
 import java.io.IOException;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import com.goodee.corpdesk.attendance.DTO.AttendanceEditDTO;
+import com.goodee.corpdesk.attendance.entity.Attendance;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellType;
 import org.apache.poi.ss.usermodel.Row;
@@ -21,17 +24,11 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-import com.goodee.corpdesk.attendance.AttendanceService;
+import com.goodee.corpdesk.attendance.service.AttendanceService;
 import com.goodee.corpdesk.employee.Employee.CreateGroup;
 import com.goodee.corpdesk.employee.Employee.UpdateGroup;
 import com.goodee.corpdesk.employee.validation.UpdateEmail;
@@ -54,6 +51,9 @@ public class EmployeeController {
 
     private final EmployeeService employeeService;
     private final AttendanceService attendanceService;
+
+    private final DateTimeFormatter formatterInput = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm");
+    private final DateTimeFormatter formatterOutput = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 
     public EmployeeController(EmployeeService employeeService, AttendanceService attendanceService, PasswordEncoder passwordEncoder) {
         this.employeeService = employeeService;
@@ -415,5 +415,103 @@ public class EmployeeController {
 		}
 
         return "redirect:/logout";
+    }
+
+    // AttendanceContoroller에 있던, url이 /employee/로 시작하는 메서드들 여기로 옮김
+    @PostMapping("{username}/attendance/delete")
+    @ResponseBody
+    public Map<String, Object> deleteAttendance(@PathVariable("username") String username,
+                                                @RequestBody Map<String, List<Long>> payload) {
+        List<Long> attendanceIds = payload.get("attendanceIds");
+        Map<String, Object> result = new HashMap<>();
+        try {
+            attendanceService.deleteAttendances(username, attendanceIds);
+            result.put("success", true);
+        } catch (Exception e) {
+            result.put("success", false);
+            result.put("error", e.getMessage());
+        }
+        return result;
+    }
+
+    // 출퇴근 추가
+    @PostMapping("{username}/attendance/add")
+    @ResponseBody
+    public Map<String, Object> addAttendance(
+        @PathVariable(name = "username") String username,   // 명시적으로 name 지정
+        @RequestBody AttendanceEditDTO dto) {
+
+        Map<String, Object> result = new HashMap<>();
+        try {
+            Attendance att = new Attendance();
+            att.setUsername(username);
+            att.setWorkStatus(dto.getWorkStatus());
+            att.setUseYn(true);
+
+            if (dto.getDateTime() != null && !dto.getDateTime().isEmpty()) {
+                LocalDateTime dt = LocalDateTime.parse(dto.getDateTime(),
+                    DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm"));
+                if ("출근".equals(dto.getWorkStatus())) {
+                    att.setCheckInDateTime(dt);
+                } else if ("퇴근".equals(dto.getWorkStatus())) {
+                    att.setCheckOutDateTime(dt);
+                }
+            }
+
+            Attendance saved = attendanceService.saveOrUpdateAttendance(att);
+            result.put("success", true);
+            result.put("attendanceId", saved.getAttendanceId());
+            result.put("workStatus", saved.getWorkStatus());
+            result.put("dateTime", saved.getCheckInDateTime() != null ?
+                saved.getCheckInDateTime().format(formatterOutput) :
+                (saved.getCheckOutDateTime() != null ?
+                    saved.getCheckOutDateTime().format(formatterOutput) : null));
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            result.put("success", false);
+            result.put("error", e.getMessage());
+        }
+
+        return result;
+
+    }
+
+
+    @PostMapping("{username}/attendance/edit")
+    @ResponseBody
+    public Map<String, Object> editAttendance(@PathVariable("username") String username,
+                                              @RequestBody AttendanceEditDTO dto) {
+
+        Map<String, Object> result = new java.util.HashMap<>();
+        try {
+            Attendance attendance = attendanceService.getAttendanceById(dto.getAttendanceId());
+
+            if (!org.springframework.util.StringUtils.hasText(dto.getWorkStatus())) {
+                attendance.setWorkStatus("-");
+            } else {
+                attendance.setWorkStatus(dto.getWorkStatus());
+            }
+
+            if (org.springframework.util.StringUtils.hasText(dto.getDateTime())) {
+                LocalDateTime dateTime = LocalDateTime.parse(dto.getDateTime(), formatterInput);
+                if ("출근".equals(attendance.getWorkStatus())) {
+                    attendance.setCheckInDateTime(dateTime);
+                } else if ("퇴근".equals(attendance.getWorkStatus())) {
+                    attendance.setCheckOutDateTime(dateTime);
+                }
+            }
+
+            attendanceService.saveOrUpdateAttendance(attendance);
+
+            result.put("success", true);
+        } catch (Exception e) {
+            e.printStackTrace();
+            result.put("success", false);
+            result.put("error", e.getMessage());
+        }
+
+        return result;
+
     }
 }
